+++ extensions/amp-analytics/0.1/test/test-config-new.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 The AMP HTML Authors. All Rights Reserved.
+ * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import {ANALYTICS_CONFIG} from '../vendors';
 import {AnalyticsConfig, expandConfigRequest, mergeObjects} from '../config';
 import {Services} from '../../../../src/services';
 import {installDocService} from '../../../../src/service/ampdoc-impl';
@@ -22,28 +21,46 @@ import {map} from '../../../../src/utils/object';
 import {stubService} from '../../../../testing/test-helper';
 import {user} from '../../../../src/log';
 
-// TODO(zhouyx@): Remove after ANALYTICS_VENDOR_SPLIT clean up
-describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
+describes.realWin('AnalyticsConfig', {amp: false}, env => {
   let win;
   let doc;
+  const vendorName = 'test-vendor';
+  let fakeVendorJson;
+  let fakeRemoteJson;
+  let fakeRewriterJson;
+  let xhrStub;
 
   beforeEach(() => {
     win = env.win;
     doc = win.document;
-  });
-
-  afterEach(() => {
-    delete ANALYTICS_CONFIG['-test-venfor'];
+    xhrStub = stubXhr();
+    xhrStub.callsFake(url => {
+      if (url.indexOf(`${vendorName}.json`) >= 0) {
+        return Promise.resolve({
+          json: () => fakeVendorJson,
+        });
+      }
+      if (url.indexOf('/remote') >= 0) {
+        return Promise.resolve({
+          json: () => fakeRemoteJson,
+        });
+      }
+      if (url.indexOf('/rewriter') >= 0) {
+        return Promise.resolve({
+          json: () => fakeRewriterJson,
+        });
+      }
+    });
   });
 
   describe('handles top level fields correctly', () => {
     it('propogates requestOrigin into each request object', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requestOrigin': 'https://example.com',
         'requests': {'test1': '/test1', 'test2': '/test1/test2'},
       };
 
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -60,7 +77,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('does not overwrite existing origin in request object', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requestOrigin': 'https://toplevel.com',
         'requests': {
           'test1': {
@@ -70,7 +87,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
         },
       };
 
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -83,7 +100,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('handles empty string request origin', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requestOrigin': '',
         'requests': {
           'test1': {
@@ -92,7 +109,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
         },
       };
 
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -105,7 +122,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('handles undefined request origin', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requestOrigin': undefined,
         'requests': {
           'test1': {
@@ -114,7 +131,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
         },
       };
 
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -129,7 +146,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
 
   describe('merges requests correctly', () => {
     it('inline and vendor both string', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '/bar', 'bar': 'foobar'},
       };
 
@@ -138,7 +155,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': 'https://example.com/${bar}'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -160,7 +177,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('inline and vendor string and object', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {
           'foo': 'foo',
           'bar': {
@@ -180,7 +197,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           },
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -204,7 +221,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('inline and vendor both object', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {
           'foo': {
             'baseUrl': 'foo',
@@ -228,7 +245,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           },
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -252,6 +269,12 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('inline and remote both string', () => {
+      fakeRemoteJson = {
+        requests: {
+          foo: 'https://example.com/remote',
+        },
+      };
+
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -259,20 +282,8 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '//config-rv2',
+          'config': '/remote',
         }
-      );
-
-      stubXhr().returns(
-        Promise.resolve({
-          json() {
-            return Promise.resolve({
-              requests: {
-                foo: 'https://example.com/remote',
-              },
-            });
-          },
-        })
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -332,7 +343,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
 
   describe('vendor only configs', () => {
     it('succeeds for vendor optout config', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -344,7 +355,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
         },
         optout: true,
       };
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['optout']).to.be.true;
       });
@@ -364,7 +375,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('succeeds for vendor iframePing config', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -376,7 +387,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           },
         },
       };
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['triggers']).to.deep.equal({
           'pageview': {
@@ -401,7 +412,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('succeeds for vendor iframe transport config', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -418,7 +429,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           iframe: '//fake-url',
         },
       };
-      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
+      const element = getAnalyticsTag({}, {'type': vendorName});
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['transport']).to.deep.equal({
           image: false,
@@ -437,7 +448,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
         '[AmpAnalytics <unknown id>] Inline or remote config ' +
           'should not overwrite vendor transport settings'
       );
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -465,7 +476,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
             iframe: '//fake-url2',
           },
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
       return new AnalyticsConfig(element).loadConfig();
     });
@@ -477,6 +488,13 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
         /Remote configs are not allowed to specify transport iframe/
       );
 
+      fakeRemoteJson = {
+        vars: {'title': 'remote'},
+        transport: {
+          iframe: '//fake-url',
+        },
+      };
+
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -484,26 +502,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '//config1',
+          'config': '/remote',
         }
       );
 
-      const xhrStub = stubXhr();
-      xhrStub.returns(
-        Promise.resolve({
-          json: () => {
-            return {
-              vars: {'title': 'remote'},
-              transport: {
-                iframe: '//fake-url',
-              },
-            };
-          },
-        })
-      );
-
       return new AnalyticsConfig(element).loadConfig().then(config => {
-        expect(xhrStub).to.be.calledWith('//config1', {});
         expect(config['vars']['title']).to.equal('remote');
         // iframe transport from remote config is ignored
         expect(config['transport']['iframe']).to.be.undefined;
@@ -511,6 +514,9 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should not fetch remote config if sandboxed', () => {
+      fakeRemoteJson = {
+        vars: {'title': 'remote'},
+      };
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -518,7 +524,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '//config1',
+          'config': '/remote',
           'sandbox': 'true',
         }
       );
@@ -529,6 +535,10 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('fetches and merges remote config with credentials', () => {
+      fakeRemoteJson = {
+        vars: {'title': 'remote'},
+      };
+
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -536,23 +546,13 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '//config1',
+          'config': '/remote',
           'data-credentials': 'include',
         }
       );
 
-      const xhrStub = stubXhr();
-      xhrStub.returns(
-        Promise.resolve({
-          json: () => {
-            return {
-              vars: {'title': 'remote'},
-            };
-          },
-        })
-      );
       return new AnalyticsConfig(element).loadConfig().then(config => {
-        expect(xhrStub).to.be.calledWith('//config1', {
+        expect(xhrStub).to.be.calledWith('/remote', {
           credentials: 'include',
         });
         expect(config['vars']['title']).to.equal('remote');
@@ -562,53 +562,38 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
 
   describe('should re-write configuration if configured', () => {
     it('should fully rewrite a configuration', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '//rewriter',
+          'url': '/rewriter',
         },
+      };
+      fakeRewriterJson = {
+        'requests': {'foo': '/rewriter'},
+        'triggers': [{'on': 'visible', 'request': 'foo'}],
       };
       const element = getAnalyticsTag(
         {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
-      const xhrStub = stubXhr();
-      xhrStub.callsFake(url => {
-        const result = {
-          'requests': {'foo': url},
-          'triggers': [{'on': 'visible', 'request': 'foo'}],
-        };
-        return Promise.resolve({
-          json: () => result,
-        });
-      });
-
       return new AnalyticsConfig(element).loadConfig().then(config => {
-        expect(xhrStub).to.be.calledWith('//rewriter', {
-          body: {
-            requests: {foo: '//inlined'},
-            triggers: [{on: 'visible', request: 'foo'}],
-          },
-          method: 'POST',
-        });
-
         expect(config['requests']['foo']).to.deep.equal({
-          baseUrl: '//rewriter',
+          baseUrl: '/rewriter',
         });
       });
     });
 
     it('should resolve and send publisher enabled varGroups', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '//rewriter',
+          'url': '/rewriter',
           'varGroups': {
             'feature1': {
               'key': 'cats',
@@ -620,6 +605,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           },
         },
       };
+
       const element = getAnalyticsTag(
         {
           'requests': {'foo': '//inlined'},
@@ -630,13 +616,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
             },
           },
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
-      const xhrStub = stubXhr();
-
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('//rewriter', {
+        expect(xhrStub).to.be.calledWith('/rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -653,11 +637,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should resolve and send vendor enabled varGroups', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '//rewriter',
+          'url': '/rewriter',
           'varGroups': {
             'feature1': {
               'key': 'cats',
@@ -672,13 +656,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
-      const xhrStub = stubXhr();
-
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('//rewriter', {
+        expect(xhrStub).to.be.calledWith('/rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -695,11 +677,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should not send configRewriter object if no vars are enabled', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '//rewriter',
+          'url': '/rewriter',
           'varGroups': {
             'feature1': {
               'key': 'cats',
@@ -713,13 +695,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
-      const xhrStub = stubXhr();
-
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('//rewriter', {
+        expect(xhrStub).to.be.calledWith('/rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -730,11 +710,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should support amp-analytics-variables macros in varGroups', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '//rewriter',
+          'url': '/rewriter',
           'varGroups': {
             'feature1': {
               'hasValue': '$NOT(foo)',
@@ -748,13 +728,11 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
-      const xhrStub = stubXhr();
-
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('//rewriter', {
+        expect(xhrStub).to.be.calledWith('/rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -770,34 +748,31 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should merge rewritten configuration and use vendor', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '//rewriter',
+          'url': '/rewriter',
         },
+      };
+      fakeRemoteJson = {
+        'requests': {'foo': '//remote'},
+      };
+
+      fakeRewriterJson = {
+        'requests': {'foo': '//rewriter'},
+        'triggers': [{'on': 'visible', 'request': 'foo'}],
       };
       const element = getAnalyticsTag(
         {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor', 'config': '//remote'}
+        {'type': vendorName, 'config': '//remote'}
       );
 
-      const xhrStub = stubXhr();
-      xhrStub.callsFake(url => {
-        const result = {
-          'requests': {'foo': url},
-          'triggers': [{'on': 'visible', 'request': 'foo'}],
-        };
-        return Promise.resolve({
-          json: () => result,
-        });
-      });
-
       return new AnalyticsConfig(element).loadConfig().then(config => {
-        expect(xhrStub).to.be.calledWith('//rewriter', {
+        expect(xhrStub).to.be.calledWith('/rewriter', {
           body: {
             requests: {foo: '//remote'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -812,7 +787,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should ignore config rewriter if no url provided', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
+      fakeVendorJson = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {},
@@ -822,7 +797,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': '-test-venfor'}
+        {'type': vendorName}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -833,20 +808,16 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should ignore inlined config rewriter', () => {
-      ANALYTICS_CONFIG['-test-venfor'] = {
-        'requests': {'foo': '//vendor'},
-        'triggers': [{'on': 'visible', 'request': 'foo'}],
+      fakeRewriterJson = {
+        'requests': {'foo': '//rewriter'},
       };
-      const element = getAnalyticsTag(
-        {
-          'requests': {'foo': '//inlined'},
-          'triggers': [{'on': 'visible', 'request': 'foo'}],
-          'configRewriter': {
-            'url': '//rewriter',
-          },
+      const element = getAnalyticsTag({
+        'requests': {'foo': '//inlined'},
+        'triggers': [{'on': 'visible', 'request': 'foo'}],
+        'configRewriter': {
+          'url': '//rewriter',
         },
-        {'type': '-test-venfor'}
-      );
+      });
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']['foo']).to.deep.equal({
@@ -884,14 +855,14 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
 
   describe('warning message', () => {
     it('shows the warning', () => {
-      ANALYTICS_CONFIG['test-vendor'] = {
+      fakeVendorJson = {
         'requests': {'test1': '/test1', 'test2': '/test1/test2'},
         'warningMessage': 'I am a warning',
       };
 
       const element = getAnalyticsTag(
         {},
-        {'type': 'test-vendor', 'id': 'analyticsId'}
+        {'type': vendorName, 'id': 'analyticsId'}
       );
       const usrObj = user();
       const spy = env.sandbox.spy(usrObj, 'warn');
@@ -910,7 +881,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('handles incorrect inputs', () => {
-      ANALYTICS_CONFIG['test-vendor'] = {
+      fakeVendorJson = {
         'requests': {'test1': '/test1', 'test2': '/test2'},
         'warningMessage': {
           'message': 'I am deprecated',
@@ -920,7 +891,7 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
 
       const element = getAnalyticsTag(
         {},
-        {'type': 'test-vendor', 'id': 'analyticsId'}
+        {'type': vendorName, 'id': 'analyticsId'}
       );
       const usrObj = user();
       const spy = env.sandbox.spy(usrObj, 'warn');
@@ -946,7 +917,6 @@ describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
 
       const usrObj = user();
       const spy = env.sandbox.spy(usrObj, 'warn');
-      const xhrStub = stubXhr();
       xhrStub.returns(
         Promise.resolve({
           json: () => {
