+++ extensions/amp-analytics/0.1/test/test-config.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
+ * Copyright 2018 The AMP HTML Authors. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import {ANALYTICS_CONFIG} from '../vendors';
 import {AnalyticsConfig, expandConfigRequest, mergeObjects} from '../config';
 import {Services} from '../../../../src/services';
 import {installDocService} from '../../../../src/service/ampdoc-impl';
@@ -21,46 +22,28 @@ import {map} from '../../../../src/utils/object';
 import {stubService} from '../../../../testing/test-helper';
 import {user} from '../../../../src/log';
 
-describes.realWin('AnalyticsConfig', {amp: false}, env => {
+// TODO(zhouyx@): Remove after ANALYTICS_VENDOR_SPLIT clean up
+describes.realWin.skip('AnalyticsConfig', {amp: false}, env => {
   let win;
   let doc;
-  const vendorName = 'test-vendor';
-  let fakeVendorJson;
-  let fakeRemoteJson;
-  let fakeRewriterJson;
-  let xhrStub;
 
   beforeEach(() => {
     win = env.win;
     doc = win.document;
-    xhrStub = stubXhr();
-    xhrStub.callsFake(url => {
-      if (url.indexOf(`${vendorName}.json`) >= 0) {
-        return Promise.resolve({
-          json: () => fakeVendorJson,
-        });
-      }
-      if (url.indexOf('/remote') >= 0) {
-        return Promise.resolve({
-          json: () => fakeRemoteJson,
-        });
-      }
-      if (url.indexOf('/rewriter') >= 0) {
-        return Promise.resolve({
-          json: () => fakeRewriterJson,
-        });
-      }
-    });
+  });
+
+  afterEach(() => {
+    delete ANALYTICS_CONFIG['-test-venfor'];
   });
 
   describe('handles top level fields correctly', () => {
     it('propogates requestOrigin into each request object', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requestOrigin': 'https://example.com',
         'requests': {'test1': '/test1', 'test2': '/test1/test2'},
       };
 
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -77,7 +60,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('does not overwrite existing origin in request object', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requestOrigin': 'https://toplevel.com',
         'requests': {
           'test1': {
@@ -87,7 +70,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
         },
       };
 
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -100,7 +83,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('handles empty string request origin', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requestOrigin': '',
         'requests': {
           'test1': {
@@ -109,7 +92,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
         },
       };
 
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -122,7 +105,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('handles undefined request origin', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requestOrigin': undefined,
         'requests': {
           'test1': {
@@ -131,7 +114,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
         },
       };
 
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']).to.deep.equal({
@@ -146,7 +129,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
 
   describe('merges requests correctly', () => {
     it('inline and vendor both string', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '/bar', 'bar': 'foobar'},
       };
 
@@ -155,7 +138,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': 'https://example.com/${bar}'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -177,7 +160,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('inline and vendor string and object', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {
           'foo': 'foo',
           'bar': {
@@ -197,7 +180,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           },
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -221,7 +204,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('inline and vendor both object', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {
           'foo': {
             'baseUrl': 'foo',
@@ -245,7 +228,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           },
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -269,12 +252,6 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('inline and remote both string', () => {
-      fakeRemoteJson = {
-        requests: {
-          foo: 'https://example.com/remote',
-        },
-      };
-
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -282,8 +259,20 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '/remote',
+          'config': '//config-rv2',
         }
+      );
+
+      stubXhr().returns(
+        Promise.resolve({
+          json() {
+            return Promise.resolve({
+              requests: {
+                foo: 'https://example.com/remote',
+              },
+            });
+          },
+        })
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -343,7 +332,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
 
   describe('vendor only configs', () => {
     it('succeeds for vendor optout config', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -355,7 +344,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
         },
         optout: true,
       };
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['optout']).to.be.true;
       });
@@ -375,7 +364,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('succeeds for vendor iframePing config', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -387,7 +376,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           },
         },
       };
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['triggers']).to.deep.equal({
           'pageview': {
@@ -412,7 +401,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('succeeds for vendor iframe transport config', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -429,7 +418,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           iframe: '//fake-url',
         },
       };
-      const element = getAnalyticsTag({}, {'type': vendorName});
+      const element = getAnalyticsTag({}, {'type': '-test-venfor'});
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['transport']).to.deep.equal({
           image: false,
@@ -448,7 +437,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
         '[AmpAnalytics <unknown id>] Inline or remote config ' +
           'should not overwrite vendor transport settings'
       );
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {
           'pageview': '//fake-url',
         },
@@ -476,7 +465,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
             iframe: '//fake-url2',
           },
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
       return new AnalyticsConfig(element).loadConfig();
     });
@@ -488,13 +477,6 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
         /Remote configs are not allowed to specify transport iframe/
       );
 
-      fakeRemoteJson = {
-        vars: {'title': 'remote'},
-        transport: {
-          iframe: '//fake-url',
-        },
-      };
-
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -502,11 +484,26 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '/remote',
+          'config': '//config1',
         }
       );
 
+      const xhrStub = stubXhr();
+      xhrStub.returns(
+        Promise.resolve({
+          json: () => {
+            return {
+              vars: {'title': 'remote'},
+              transport: {
+                iframe: '//fake-url',
+              },
+            };
+          },
+        })
+      );
+
       return new AnalyticsConfig(element).loadConfig().then(config => {
+        expect(xhrStub).to.be.calledWith('//config1', {});
         expect(config['vars']['title']).to.equal('remote');
         // iframe transport from remote config is ignored
         expect(config['transport']['iframe']).to.be.undefined;
@@ -514,9 +511,6 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should not fetch remote config if sandboxed', () => {
-      fakeRemoteJson = {
-        vars: {'title': 'remote'},
-      };
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -524,7 +518,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '/remote',
+          'config': '//config1',
           'sandbox': 'true',
         }
       );
@@ -535,10 +529,6 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('fetches and merges remote config with credentials', () => {
-      fakeRemoteJson = {
-        vars: {'title': 'remote'},
-      };
-
       const element = getAnalyticsTag(
         {
           'vars': {'title': 'local'},
@@ -546,13 +536,23 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
         {
-          'config': '/remote',
+          'config': '//config1',
           'data-credentials': 'include',
         }
       );
 
+      const xhrStub = stubXhr();
+      xhrStub.returns(
+        Promise.resolve({
+          json: () => {
+            return {
+              vars: {'title': 'remote'},
+            };
+          },
+        })
+      );
       return new AnalyticsConfig(element).loadConfig().then(config => {
-        expect(xhrStub).to.be.calledWith('/remote', {
+        expect(xhrStub).to.be.calledWith('//config1', {
           credentials: 'include',
         });
         expect(config['vars']['title']).to.equal('remote');
@@ -562,38 +562,53 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
 
   describe('should re-write configuration if configured', () => {
     it('should fully rewrite a configuration', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '/rewriter',
+          'url': '//rewriter',
         },
-      };
-      fakeRewriterJson = {
-        'requests': {'foo': '/rewriter'},
-        'triggers': [{'on': 'visible', 'request': 'foo'}],
       };
       const element = getAnalyticsTag(
         {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
+      const xhrStub = stubXhr();
+      xhrStub.callsFake(url => {
+        const result = {
+          'requests': {'foo': url},
+          'triggers': [{'on': 'visible', 'request': 'foo'}],
+        };
+        return Promise.resolve({
+          json: () => result,
+        });
+      });
+
       return new AnalyticsConfig(element).loadConfig().then(config => {
+        expect(xhrStub).to.be.calledWith('//rewriter', {
+          body: {
+            requests: {foo: '//inlined'},
+            triggers: [{on: 'visible', request: 'foo'}],
+          },
+          method: 'POST',
+        });
+
         expect(config['requests']['foo']).to.deep.equal({
-          baseUrl: '/rewriter',
+          baseUrl: '//rewriter',
         });
       });
     });
 
     it('should resolve and send publisher enabled varGroups', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '/rewriter',
+          'url': '//rewriter',
           'varGroups': {
             'feature1': {
               'key': 'cats',
@@ -605,7 +620,6 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           },
         },
       };
-
       const element = getAnalyticsTag(
         {
           'requests': {'foo': '//inlined'},
@@ -616,11 +630,13 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
             },
           },
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
+      const xhrStub = stubXhr();
+
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('/rewriter', {
+        expect(xhrStub).to.be.calledWith('//rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -637,11 +653,11 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should resolve and send vendor enabled varGroups', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '/rewriter',
+          'url': '//rewriter',
           'varGroups': {
             'feature1': {
               'key': 'cats',
@@ -656,11 +672,13 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
+      const xhrStub = stubXhr();
+
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('/rewriter', {
+        expect(xhrStub).to.be.calledWith('//rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -677,11 +695,11 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should not send configRewriter object if no vars are enabled', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '/rewriter',
+          'url': '//rewriter',
           'varGroups': {
             'feature1': {
               'key': 'cats',
@@ -695,11 +713,13 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
+      const xhrStub = stubXhr();
+
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('/rewriter', {
+        expect(xhrStub).to.be.calledWith('//rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -710,11 +730,11 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should support amp-analytics-variables macros in varGroups', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '/rewriter',
+          'url': '//rewriter',
           'varGroups': {
             'feature1': {
               'hasValue': '$NOT(foo)',
@@ -728,11 +748,13 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
+      const xhrStub = stubXhr();
+
       return new AnalyticsConfig(element).loadConfig().then(() => {
-        expect(xhrStub).to.be.calledWith('/rewriter', {
+        expect(xhrStub).to.be.calledWith('//rewriter', {
           body: {
             requests: {foo: '//inlined'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -748,31 +770,34 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should merge rewritten configuration and use vendor', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {
-          'url': '/rewriter',
+          'url': '//rewriter',
         },
-      };
-      fakeRemoteJson = {
-        'requests': {'foo': '//remote'},
-      };
-
-      fakeRewriterJson = {
-        'requests': {'foo': '//rewriter'},
-        'triggers': [{'on': 'visible', 'request': 'foo'}],
       };
       const element = getAnalyticsTag(
         {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName, 'config': '//remote'}
+        {'type': '-test-venfor', 'config': '//remote'}
       );
 
+      const xhrStub = stubXhr();
+      xhrStub.callsFake(url => {
+        const result = {
+          'requests': {'foo': url},
+          'triggers': [{'on': 'visible', 'request': 'foo'}],
+        };
+        return Promise.resolve({
+          json: () => result,
+        });
+      });
+
       return new AnalyticsConfig(element).loadConfig().then(config => {
-        expect(xhrStub).to.be.calledWith('/rewriter', {
+        expect(xhrStub).to.be.calledWith('//rewriter', {
           body: {
             requests: {foo: '//remote'},
             triggers: [{on: 'visible', request: 'foo'}],
@@ -787,7 +812,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should ignore config rewriter if no url provided', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['-test-venfor'] = {
         'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
         'configRewriter': {},
@@ -797,7 +822,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
           'requests': {'foo': '//inlined'},
           'triggers': [{'on': 'visible', 'request': 'foo'}],
         },
-        {'type': vendorName}
+        {'type': '-test-venfor'}
       );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
@@ -808,16 +833,20 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('should ignore inlined config rewriter', () => {
-      fakeRewriterJson = {
-        'requests': {'foo': '//rewriter'},
-      };
-      const element = getAnalyticsTag({
-        'requests': {'foo': '//inlined'},
+      ANALYTICS_CONFIG['-test-venfor'] = {
+        'requests': {'foo': '//vendor'},
         'triggers': [{'on': 'visible', 'request': 'foo'}],
-        'configRewriter': {
-          'url': '//rewriter',
+      };
+      const element = getAnalyticsTag(
+        {
+          'requests': {'foo': '//inlined'},
+          'triggers': [{'on': 'visible', 'request': 'foo'}],
+          'configRewriter': {
+            'url': '//rewriter',
+          },
         },
-      });
+        {'type': '-test-venfor'}
+      );
 
       return new AnalyticsConfig(element).loadConfig().then(config => {
         expect(config['requests']['foo']).to.deep.equal({
@@ -855,14 +884,14 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
 
   describe('warning message', () => {
     it('shows the warning', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['test-vendor'] = {
         'requests': {'test1': '/test1', 'test2': '/test1/test2'},
         'warningMessage': 'I am a warning',
       };
 
       const element = getAnalyticsTag(
         {},
-        {'type': vendorName, 'id': 'analyticsId'}
+        {'type': 'test-vendor', 'id': 'analyticsId'}
       );
       const usrObj = user();
       const spy = env.sandbox.spy(usrObj, 'warn');
@@ -881,7 +910,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
     });
 
     it('handles incorrect inputs', () => {
-      fakeVendorJson = {
+      ANALYTICS_CONFIG['test-vendor'] = {
         'requests': {'test1': '/test1', 'test2': '/test2'},
         'warningMessage': {
           'message': 'I am deprecated',
@@ -891,7 +920,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
 
       const element = getAnalyticsTag(
         {},
-        {'type': vendorName, 'id': 'analyticsId'}
+        {'type': 'test-vendor', 'id': 'analyticsId'}
       );
       const usrObj = user();
       const spy = env.sandbox.spy(usrObj, 'warn');
@@ -917,6 +946,7 @@ describes.realWin('AnalyticsConfig', {amp: false}, env => {
 
       const usrObj = user();
       const spy = env.sandbox.spy(usrObj, 'warn');
+      const xhrStub = stubXhr();
       xhrStub.returns(
         Promise.resolve({
           json: () => {
