+++ extensions/amp-analytics/0.1/test/test-vendors-new.js
@@ -1,0 +1,208 @@
+/**
+ * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {AmpAnalytics} from '../amp-analytics';
+import {AnalyticsConfig} from '../config';
+import {ExpansionOptions} from '../variables';
+import {IFRAME_TRANSPORTS} from '../iframe-transport-vendors';
+import {
+  ImagePixelVerifier,
+  mockWindowInterface,
+} from '../../../../testing/test-helper';
+import {Services} from '../../../../src/services';
+import {hasOwn} from '../../../../src/utils/object';
+import {macroTask} from '../../../../testing/yield';
+
+/* global require: false */
+const VENDOR_REQUESTS = require('./vendor-requests.json');
+
+describes.realWin(
+  'amp-analytics',
+  {
+    amp: {
+      extensions: ['amp-analytics'],
+    },
+    mockFetch: false,
+  },
+  function(env) {
+    let win, doc;
+    let requestVerifier;
+
+    beforeEach(() => {
+      win = env.win;
+      doc = win.document;
+      const wi = mockWindowInterface(env.sandbox);
+      wi.getLocation.returns(win.location);
+      requestVerifier = new ImagePixelVerifier(wi);
+    });
+
+    describe('Should not contain iframe transport if not whitelisted', () => {
+      for (const vendor in VENDOR_REQUESTS) {
+        it('test vendor: ' + vendor, () => {
+          const el = doc.createElement('amp-analytics');
+          el.setAttribute('type', vendor);
+          doc.body.appendChild(el);
+          const analyticsConfig = new AnalyticsConfig(el);
+          return analyticsConfig.loadConfig().then(config => {
+            if (
+              hasOwn(config, 'transport') &&
+              hasOwn(config.transport, 'iframe')
+            ) {
+              expect(config['transport']['iframe']).to.equal(
+                IFRAME_TRANSPORTS[vendor]
+              );
+            }
+          });
+        });
+      }
+    });
+
+    describe('vendor request tests', () => {
+      for (const vendor in VENDOR_REQUESTS) {
+        describe('analytics vendor: ' + vendor, function() {
+          let config;
+          let analytics;
+          beforeEach(done => {
+            const urlReplacements = Services.urlReplacementsForDoc(
+              doc.documentElement
+            );
+            window.sandbox
+              .stub(urlReplacements.getVariableSource(), 'get')
+              .callsFake(function(name) {
+                expect(this.replacements_).to.have.property(name);
+                const defaultValue = `_${name.toLowerCase()}_`;
+                return {
+                  sync: () => defaultValue,
+                };
+              });
+
+            window.sandbox
+              .stub(ExpansionOptions.prototype, 'getVar')
+              .callsFake(function(name) {
+                let val = this.vars[name];
+                if (val == null || val == '') {
+                  val = '!' + name;
+                }
+                return val;
+              });
+
+            analytics = getAnalyticsTag(doc, vendor);
+            analytics.layoutCallback().then(() => {
+              config = analytics.config_;
+              done();
+            });
+          });
+
+          it('test requests', function*() {
+            const outputConfig = {};
+            for (const name in config.requests) {
+              // Wait for event queue to clear.
+              // To prevent default triggers
+              yield macroTask();
+              analytics.handleEvent_(
+                {
+                  request: name,
+                },
+                {
+                  vars: Object.create(null),
+                }
+              );
+
+              // Reset $requestCount value after each
+              analytics.config_['vars']['requestCount'] = 0;
+
+              yield macroTask();
+              expect(requestVerifier.hasRequestSent()).to.be.true;
+              let url = requestVerifier.getLastRequestUrl();
+
+              const vendorData = VENDOR_REQUESTS[vendor];
+              if (!vendorData) {
+                throw new Error(
+                  'Add vendor ' + vendor + ' to vendor-requests.json'
+                );
+              }
+              const val = vendorData[name];
+              if (val == '<ignore for test>') {
+                url = '<ignore for test>';
+              }
+              if (val == null) {
+                throw new Error(
+                  'Define ' +
+                    vendor +
+                    '.' +
+                    name +
+                    ' in vendor-requests.json. Expected value: ' +
+                    url
+                );
+              }
+              outputConfig[name] = url;
+              // Write this out for easy copy pasting.
+              if (url !== val) {
+                throw new Error(
+                  `Vendor ${vendor}, request ${name} doesn't match` +
+                    `Expected value ${val}, get value ${url}`
+                );
+              }
+              expect(url).to.equal(val);
+            }
+            writeOutput(vendor, outputConfig);
+          });
+        });
+      }
+    });
+  }
+);
+
+/**
+ * Get the AmpAnalytics instance
+ * @param {string} vendor
+ */
+function getAnalyticsTag(doc, vendor) {
+  const el = doc.createElement('amp-analytics');
+  el.setAttribute('type', vendor);
+
+  // Overwrite transport method to use image
+  expectAsyncConsoleError(/should not overwrite vendor transport/);
+  const script = doc.createElement('script');
+  script.setAttribute('type', 'application/json');
+  script.textContent = JSON.stringify({
+    'transport': {
+      'iframe': false,
+      'image': true,
+    },
+  });
+  el.appendChild(script);
+
+  doc.body.appendChild(el);
+  el.connectedCallback();
+  const analytics = new AmpAnalytics(el);
+  analytics.createdCallback();
+  analytics.buildCallback();
+  return analytics;
+}
+
+// Cache firstChild before append div to doc
+const {firstChild} = top.document.body;
+
+function writeOutput(vendor, output) {
+  const vendorDiv = top.document.createElement('h3');
+  vendorDiv.textContent = vendor;
+  top.document.body.insertBefore(vendorDiv, firstChild);
+
+  const out = top.document.createElement('div');
+  out.textContent = JSON.stringify(output, null, '  ');
+  top.document.body.insertBefore(out, firstChild);
+}
